# api-spec.yaml
openapi: 3.0.3
info:
  title: Task Manager API
  version: "1.0.0"
servers:
  - url: /api
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [status, timestamp]
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time
  /tasks:
    get:
      summary: List tasks
      parameters:
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - in: query
          name: priority
          schema:
            $ref: "#/components/schemas/TaskPriority"
        - in: query
          name: search
          description: Case-insensitive substring match on title
          schema:
            type: string
            maxLength: 255
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: object
                required: [data, count]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
                  count:
                    type: integer
                    minimum: 0
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
  /tasks/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get a task by id
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      summary: Partially update a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchTaskRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a task
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [message, id]
                properties:
                  message:
                    type: string
                    example: Task deleted
                  id:
                    type: string
                    format: uuid
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /tasks/clear/done:
    delete:
      summary: Delete all tasks with status=done
      responses:
        "200":
          description: Deleted count
          content:
            application/json:
              schema:
                type: object
                required: [message, deleted]
                properties:
                  message:
                    type: string
                    example: Done tasks cleared
                  deleted:
                    type: integer
                    minimum: 0
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalid_uuid:
              value: { "error": "invalid id \"not-a-uuid\": expected UUID" }
            validation:
              value: { "error": "title is required and must be under 255 characters" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Task not found
  schemas:
    TaskStatus:
      type: string
      enum: [pending, in_progress, done]
    TaskPriority:
      type: string
      enum: [low, medium, high]
    Task:
      type: object
      required:
        - id
        - title
        - status
        - priority
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        due_date:
          type: string
          format: date
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTaskRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        due_date:
          type: string
          format: date
          nullable: true
    PatchTaskRequest:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        due_date:
          type: string
          format: date
          nullable: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string